ai-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
openai==1.10.0
langchain==0.1.0
langchain-openai==0.0.3
faiss-cpu
scikit-learn
pandas
numpy
xgboost==2.0.3
httpx==0.26.0
redis==5.0.1
